requirements.txt
streamlit
langchain
langchain-community
chromadb
sentence-transformers
python-dotenv
pypdf
requests
pydantic
ddgs
readability-lxml
